v1/galion/services/voice-service/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Audio Processing
pydub==0.25.1
soundfile==0.12.1
numpy==1.26.3
scipy==1.12.0

# AI & ML
openai==1.10.0
elevenlabs==0.2.26
httpx==0.26.0

# Database & Cache
sqlalchemy==2.0.25
asyncpg==0.29.0
redis==5.0.1
psycopg2-binary==2.9.9

# Event Streaming
aiokafka==0.10.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0

